+++ setup.cfg
@@ -1,3 +1,26 @@
+[metadata]
+name = portal-visualization
+version = 0.0.1
+author = Chuck McCallum
+author_email = mccallucc@gmail.com
+description = Given HuBMAP Dataset JSON, creates a Vitessce configuration.
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/hubmapconsortium/portal-visualization
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
+
 [flake8]
 max-line-length = 99
 ignore =
